KS-1-3 Fixed types of variables

+++ src/main/kotlin/lesson_1/lesson1_task3.kt
@@ -2,8 +2,8 @@ package org.example.lesson_1
 
 fun main(){
     val year: Short = 1961
-    var hour: Short = 9
-    var minute: Short = 7
+    var hour: Byte = 9
+    var minute: Byte = 7
     println(year)
     println(hour)
     println(minute)
